requirements.txt
bcrypt==3.2.0
bson==0.5.10
cachelib==0.3.0
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
cryptography==3.4.8
dnspython==2.1.0
email-validator==1.1.3
extended==0.0.8
Flask==2.0.1
Flask-JWT-Extended==4.3.0
Flask-Login==0.5.0
flask-mongoengine==1.0.0
Flask-OAuthlib==0.9.6
Flask-PyMongo==2.3.0
Flask-WTF==0.15.1
google-auth==2.0.2
google-auth-oauthlib==0.4.6
gunicorn==20.1.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
jwt==1.2.0
lib==3.0.0
MarkupSafe==2.0.1
mongoengine==0.23.1
oauth==1.0.1
oauthlib==2.1.0
pipenv==2018.11.26
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20

pymongo==3.12.0
pyOpenSSL==20.0.1
python-dateutil==2.8.2
requests==2.26.0
requests-oauthlib==1.1.0
rsa==4.7.2
six==1.16.0
urllib3==1.26.6
virtualenv==16.7.7
virtualenv-clone==0.5.3
Werkzeug==2.0.1
WTForms==2.3.3

# please install another version PyJWT to run from windows
